+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * :hammer: Premise's checkInRange and checkNotInRange method rename to requireInRange and requireNotInRange
 * :sparkles: Premise add check and checkNotNull method
 * :sparkles: Premise add requireSafe and requireNotSafe method
+* :sparkles: Aes.kt add toAesKeyFromBytes and toAesKeyFromBase64 method
 
 ## 0.6.4
 * Char add isNotBlank, isBlankOr, isNotChinese, isNotChineseOr, isDigit, isNotDigit, isNotDigitOr,